--- pages/payment/failed.tsx
@@ -1,5 +1,0 @@
-const Failed = () => {
-   return <h1>결제 실패 페이지</h1>;
-};
-
-export default Failed;
