+++ pages/payments/success.tsx
@@ -16,21 +16,26 @@ const Succeess = () => {
 
       // GET /payments/success?paymentKey=PAYMENT_KEY&orderId=ORDER_ID&amount=10000
       axios
-         .get(API_URI + `/payments/success?paymentKey=${paymentKey}&orderId=${orderId}&amount=${amount}`)
+         .get(
+            `https://27c6-39-116-11-157.jp.ngrok.io/payments/success?paymentKey=${paymentKey}&orderId=${orderId}&amount=${amount}`
+         )
          .then(res => {
             if (res.status == 200 && res.data.method === '카드') {
                window.alert({
                   position: 'center',
                   icon: 'success',
-                  text: '결제가 완료되었습니다!',
+                  text: '예약이 완료되었습니다!',
                   showConfirmButton: false,
                   timer: 1500
                });
                route.push('/attending');
             }
+
+            console.log('성공 페이지에서 : ', res);
          })
-         .catch(() => {
-            route.push('/payment/failed');
+         .catch(err => {
+            console.log('성공 페이지에서 에러 : ', err);
+            // route.push('/payment/failed');
          });
    }, []);
    return <h1>결제 성공 페이지</h1>;
